requirements.txt
anndata==0.11.4
h5py==3.14.0
igraph==0.11.9
leidenalg==0.10.2
matplotlib==3.10.6
mudata==0.3.2
numpy==2.1.2
pandas==2.3.3
scanpy==1.11.4
scikit-learn==1.7.2
scipy==1.15.3
scvi-tools==1.3.3
torch==2.8.0
tqdm==4.67.1
umap-learn==0.5.9.post2